Refactor handling of runtime configuration to prepare for reloading
